# project/api/openapi.yaml
openapi: '3.0.3'
info:
  title: 'Dynamic Data Management System API'
  description: 'API for managing dynamic entities, fields, records, and relationships.'
  version: '1.0.0'
servers:
  - url: '/api/v1'
    description: 'API v1'
...
paths:
...
  /entities/{entityTypeKey}:
    post:
      tags:
        - Records
      summary: 'Create Record'
      description: 'Creates a new record for a given entity type.'
      operationId: 'createRecord'
      parameters:
        - $ref: '#/components/parameters/entityTypeKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCreate'
      responses:
        '201':
          description: 'Record created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /entities/{entityTypeKey}/search:
    post:
      tags:
        - Records
      summary: 'Search Records'
      description: 'Searches for records of a given entity type using a filter.'
      operationId: 'searchRecords'
      parameters:
        - $ref: '#/components/parameters/entityTypeKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: 'A list of matching records.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /entities/{entityTypeKey}/{recordId}:
    get:
      tags:
        - Records
      summary: 'Get Record by ID'
      description: 'Retrieves a single record by its ID.'
      operationId: 'getRecord'
      parameters:
        - $ref: '#/components/parameters/entityTypeKey'
        - $ref: '#/components/parameters/recordId'
      responses:
        '200':
          description: 'The requested record.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Records
      summary: 'Update Record'
      description: 'Updates an existing record. Uses optimistic locking via a version number.'
      operationId: 'updateRecord'
      parameters:
        - $ref: '#/components/parameters/entityTypeKey'
        - $ref: '#/components/parameters/recordId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordUpdate'
      responses:
        '200':
          description: 'Record updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
...
    entityTypeKey:
      name: entityTypeKey
      in: path
      required: true
      description: 'The unique machine-readable key for the entity type (e.g., "user", "project").'
      schema:
        type: string
    recordId:
      name: recordId
      in: path
      required: true
      description: 'The ID of the record.'
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 'Bad Request - The request body is invalid.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Not Found - The requested resource does not exist.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflict - The request could not be completed due to a conflict with the current state of the resource, likely an optimistic locking failure.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Record:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        entityTypeId:
          type: string
          format: uuid
          readOnly: true
        version:
          type: integer
          description: 'The version number of the record, used for optimistic locking.'
          readOnly: true
        data:
          type: object
          additionalProperties: true
          description: 'A JSON object containing the custom field data for the record.'
          example:
            name: 'AI Roadmap Q4'
            status: 'Active'
            budget: 12000
        createdBy:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        updatedBy:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - entityTypeId
        - version
        - data
        - createdAt
        - updatedAt
    RecordCreate:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
          description: 'A JSON object containing the custom field data for the new record.'
      required:
        - data
    RecordUpdate:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
          description: 'A JSON object containing the fields to update.'
        version:
          type: integer
          description: 'The current version of the record being updated, for optimistic locking.'
      required:
        - data
        - version

    # Relation Schemas
...
    SearchRequest:
      type: object
      properties:
        filter:
          $ref: '#/components/schemas/Filter'
        sort:
          type: array
          items:
            $ref: '#/components/schemas/Sort'
          description: 'An array of sort objects to apply to the results.'
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
          description: 'The maximum number of records to return.'
        cursor:
          type: string
          nullable: true
          description: 'A cursor for pagination, obtained from a previous search response.'
    SearchResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/Record'
        nextCursor:
          type: string
          nullable: true
          description: 'A cursor to retrieve the next page of results. Null if there are no more results.'
        total:
          type: integer
          description: 'The total number of records matching the filter.'
      required:
        - rows
        - total
    Sort:
      type: object
      properties:
        field:
          type: string
          description: 'The key of the field to sort by (e.g., "budget", "createdAt").'
        dir:
          type: string
          enum: [asc, desc]
          default: asc
      required:
        - field
    Filter:
      type: object
      description: |
        A flexible filter object to query records. It can be a logical operator (`and`, `or`, `not`) or a field-level comparison.
        See the manifest for the full grammar.
      properties:
        op:
          type: string
          description: 'The operator for the filter.'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/Filter'
          description: 'Used with logical operators `and` and `or`.'
        field:
          type: string
          description: 'The key of the field to compare.'
        value:
          description: 'The value to compare against.'
        values:
          type: array
          description: 'An array of values for the `in` operator.'
        query:
          type: string
          description: 'The search query for `fulltext` search.'
      example:
        op: 'and'
        filters:
          - { op: 'eq', field: 'status', value: 'Active' }
          - { op: 'gte', field: 'budget', value: 10000 }
          - { op: 'fulltext', query: 'ai roadmap' }

    # Error Schema
    Error:
      type: object
      properties:
        code:
          type: string
          description: 'A machine-readable error code.'
          example: 'VALIDATION_ERROR'
        message:
          type: string
          description: 'A human-readable error message.'
          example: 'The request body is invalid.'
        requestId:
          type: string
          format: uuid
          description: 'A unique identifier for the request, useful for tracing.'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
      required:
        - code
        - message
    ValidationErrorDetail:
      type: object
      properties:
        path:
          type: string
          description: 'The path to the invalid field in the request body.'
          example: 'budget'
        code:
          type: string
          description: 'A machine-readable code for the specific validation error.'
          example: 'too_small'
        message:
          type: string
          description: 'A human-readable message for the specific validation error.'
          example: 'The budget must be at least 0.'
      required:
        - path
        - code
        - message